requirements.txt
Cython==3.0.10
asgiref==3.8.1
attr==0.3.2
Authlib==1.3.0
Babel==2.15.0
blinker==1.8.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.7
Cython==3.0.10
diff==2023.12.6
dl==0.1.0
docutils==0.21.2
email_validator==2.1.1
flake8==7.0.0
flask_babel==4.0.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.2
keyring==25.2.1
lockfile==0.12.2
mock==5.1.0
multiprocess==0.70.16
mypy_extensions==1.0.0
numpy==1.26.4
ordereddict==1.1
Pillow==10.3.0
pip_api==0.0.33
pipreqs==0.5.0
pox==0.3.4
protobuf==5.27.0
psycopg2-binary==2.9.9
pyenchant==3.2.2
pylama==8.4.1
pyOpenSSL==24.1.0
pytest==8.2.1
python-dotenv==1.0.1
railroad==0.5.0
requirementslib==3.0.0
Sphinx==7.3.7
sympy==1.12
thread==2.0.1
tornado==6.4
truststore==0.9.1
urllib3_secure_extra==0.1.0
watchdog==4.0.1
